Feat(`AddLangParameter.php`): add methods to validate language
Validate language by checking the `lang` parameter.
If `lang` parameter receive an invalid value, the value will be replaced by the browser language;
If `lang` parameter receive an invalid value and the value of browser language is not in `$idioms` array, the value will be replaced by `en`;
If `lang` parameter receive a route without language, the language will be added to the rout;
And if the `lang` parameter receive an invalid value that cannot be used in a valid route,  the user will be redirected to the main page.